Create 1) 31-05-25.py
You are working on a real-time inventory management system where you need to maintain a list of product IDs in a linked list. As new products are added to the inventory, their IDs are appended at the end of the list. You have been tasked with writing a program to handle this operation.

Write a program to insert a new product ID at the end of the linked list. After the completion of insertion, print the entire linked list to confirm that the new ID has been added.

Input Format

The first line contains an integer N, the number of product IDs to be initially added to the linked list.
The second line contains N space-separated integers, representing the product IDs.
Constraints

NA

Output Format

After inserting the product ID, print the current linked list, with elements separated by a space.